разделение файла на модули
